refator: brutal cleanup
